+++ CHANGES.md
@@ -9,6 +9,10 @@ This is a history of changes to [Remix](https://remix.run).
 - Fixed code splitting causing errors with shared chunks in development
 - Fixed dev server crash when renaming/deleting CSS files (#11)
 
+### Breaking Changes
+
+- Default to development mode when running the server (#10)
+
 ## 0.3.0 - Tue Oct 13 2020
 
 ### Features